Move types from pacemaker to seperate file
