+++ pacemaker_metrics.go
@@ -1,0 +1,56 @@
+package main
+
+// this types are for reading pacemaker configuration xml when running crm_mon command
+// and lookup the corrispective value
+type crmMon struct {
+	Version string  `xml:"version,attr"`
+	Summary summary `xml:"summary"`
+	Nodes   nodes   `xml:"nodes"`
+}
+
+type summary struct {
+	Nodes struct {
+		Number int `xml:"number,attr"`
+	} `xml:"nodes_configured"`
+	Resources resourcesConfigured `xml:"resources_configured"`
+}
+
+type resourcesConfigured struct {
+	Number   int `xml:"number,attr"`
+	Disabled int `xml:"disabled,attr"`
+	Blocked  int `xml:"blocked,attr"`
+}
+
+type nodes struct {
+	Node []node `xml:"node"`
+}
+
+type node struct {
+	Name             string     `xml:"name,attr"`
+	ID               string     `xml:"id,attr"`
+	Online           bool       `xml:"online,attr"`
+	Standby          bool       `xml:"standby,attr"`
+	StandbyOnFail    bool       `xml:"standby_onfail,attr"`
+	Maintenance      bool       `xml:"maintenance,attr"`
+	Pending          bool       `xml:"pending,attr"`
+	Unclean          bool       `xml:"unclean,attr"`
+	Shutdown         bool       `xml:"shutdown,attr"`
+	ExpectedUp       bool       `xml:"expected_up,attr"`
+	DC               bool       `xml:"is_dc,attr"`
+	ResourcesRunning int        `xml:"resources_running,attr"`
+	Type             string     `xml:"type,attr"`
+	Resources        []resource `xml:"resource"`
+}
+
+type resource struct {
+	ID             string `xml:"id,attr"`
+	Agent          string `xml:"resource_agent,attr"`
+	Role           string `xml:"role,attr"`
+	Active         bool   `xml:"active,attr"`
+	Orphaned       bool   `xml:"orphaned,attr"`
+	Blocked        bool   `xml:"blocked,attr"`
+	Managed        bool   `xml:"managed,attr"`
+	Failed         bool   `xml:"failed,attr"`
+	FailureIgnored bool   `xml:"failure_ignored,attr"`
+	NodesRunningOn int    `xml:"nodes_running_on,attr"`
+}
